chore: auto infer peer dep on release
