+++ .eslintrc
@@ -13,6 +13,8 @@
   },
   "rules": {
     "@typescript-eslint/ban-types": ["off"],
+    "@typescript-eslint/no-var-requires": ["off"],
+    "@typescript-eslint/explicit-module-boundary-types": ["off"],
     "@typescript-eslint/naming-convention": [
       "error",
       {
@@ -24,5 +26,21 @@
         }
       }
     ]
-  }
+  },
+  "overrides": [
+    {
+      // enable the rule specifically for TypeScript files
+      "files": ["*.ts", "*.tsx"],
+      "rules": {
+        "@typescript-eslint/explicit-function-return-type": ["error"]
+      }
+    },
+    {
+      // enable the rule specifically for TypeScript files
+      "files": ["*.ts", "*.tsx"],
+      "rules": {
+        "@typescript-eslint/no-var-requires": ["error"]
+      }
+    }
+  ]
 }
