+++ package.json
@@ -56,6 +56,7 @@
     "@aws-cdk/aws-s3": "^1.74.0",
     "@aws-cdk/aws-ssm": "^1.74.0",
     "@aws-cdk/core": "^1.74.0",
+    "aws-cdk": "^1.74.0",
     "aws-sdk": "^2.798.0",
     "busboy": "^0.3.1",
     "change-case": "^4.1.1",
