+++ packages/aws-sns-subscriptions/package.json
@@ -7,8 +7,9 @@
     "type": "git",
     "url": "https://github.com/NextFaze/cdk-extensions/tree/master/packages/aws-sns-subscriptions"
   },
-  "module": "ES6",
-  "peerDependencies": {},
+  "peerDependencies": {
+    "@aws-cdk/core": "0.0.0-AWS_CDK_PEER_VERSION"
+  },
   "dependencies": {},
   "keywords": [
     "aws-cdk",
